form_builder/requirements.txt
Django==5.0
django-crispy-forms==2.1
python-dotenv==1.0.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
Pillow==10.1.0
